setting up passport google oauth2

+++ controllers/api/v1/posts_api.js
@@ -1,7 +1,8 @@
-const Post=require('../../../models/post');
-const Comment=require('../../../models/comment');
+const Post = require('../../../models/post');
+const Comment = require('../../../models/comment');
+module.exports.index = async function(req, res){
 
-module.exports.index=async function(req,res){
+
     let posts = await Post.find({})
         .sort('-createdAt')
         .populate('user')
@@ -10,40 +11,41 @@ module.exports.index=async function(req,res){
             populate: {
                 path: 'user'
             }
-        })
+        });
 
-
-
-    return res.json(200,{
-        message:"list of posts",
-        posts:posts
+    return res.json(200, {
+        message: "List of posts",
+        posts: posts
     })
 }
+
 
 module.exports.destroy = async function(req, res){
 
     try{
         let post = await Post.findById(req.params.id);
 
-       // if (post.user == req.user.id){
+        if (post.user == req.user.id){
             post.remove();
 
             await Comment.deleteMany({post: req.params.id});
 
 
-
-            return res.json(200,{
-                message:"post and associated comments deleted"
-            })
-        // }else{
-        //     req.flash('error', 'You cannot delete this post!');
-        //     return res.redirect('back');
-        
+    
+            return res.json(200, {
+                message: "Post and associated comments deleted successfully!"
+            });
+        }else{
+            return res.json(401, {
+                message: "You cannot delete this post!"
+            });
+        }
 
     }catch(err){
-        
-        return res.json(500,{
-            message:"Internal server error"
+        console.log('********', err);
+        return res.json(500, {
+            message: "Internal Server Error"
         });
-    }}
+    }
     
+}
